Add passing SQLite tests

+++ test/EntityFramework.CrossStore.FunctionalTests/TestModels/SimpleEntity.cs
@@ -8,8 +8,7 @@ namespace Microsoft.Data.Entity.FunctionalTests.TestModels
         public static string ShadowPropertyName = "ShadowStringProperty";
         public static string ShadowPartitionIdName = "ShadowPartitionIdProperty";
 
-        // TODO: Change to int when SqLite supports generating int values
-        public virtual long Id { get; set; }
+        public virtual int Id { get; set; }
 
         public virtual string StringProperty { get; set; }
     }
